pyproject.toml
[tool.poetry]
name = "student-help-bot"
version = "0.1.0"
authors = ["BYTEZx0 <BYTEZx0@gmail.com>"]
description = "telegram bot students to get model question paper"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
supabase = "2.4.3"
annotated-types = "0.6.0"
anyio = "4.3.0"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
deprecation = "2.1.0"
gotrue = "2.4.2"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.0"
idna = "3.7"
packaging = "24.0"
postgrest = "0.16.3"
pydantic = "2.7.0"
pydantic-core = "2.18.1"
pytelegrambotapi = "4.17.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
realtime = "1.0.4"
requests = "2.31.0"
six = "1.16.0"
sniffio = "1.3.1"
storage3 = "0.7.4"
strenum = "0.4.15"
supafunc = "0.4.5"
typing-extensions = "4.11.0"
urllib3 = "2.2.1"
websockets = "12.0"
